Added support for OP_CONSTANT_LONG

+++ clox/value.h
@@ -6,8 +6,8 @@
 typedef double Value;
 
 typedef struct {
-    int capacity;
-    int count;
+    long capacity;
+    long count;
     Value* values;
 } ValueArray;
